refactor : Move directory in adapter package
