+++ writer-service/src/main/java/com/dasd412/api/writerservice/adapter/in/security/filter/JwtAuthenticationFilter.java
@@ -4,7 +4,7 @@ package com.dasd412.api.writerservice.adapter.in.security.filter;
 import com.dasd412.api.writerservice.adapter.in.security.auth.PrincipalDetails;
 import com.dasd412.api.writerservice.adapter.in.security.exception.LoginBadRequestException;
 import com.dasd412.api.writerservice.adapter.in.security.dto.LoginRequestDTO;
-import com.dasd412.api.writerservice.adapter.in.security.jwt.JWTTokenProvider;
+import com.dasd412.api.writerservice.adapter.in.security.JWTTokenProvider;
 import com.dasd412.api.writerservice.adapter.out.web.cookie.CookieProvider;
 import com.dasd412.api.writerservice.application.service.security.refresh.RefreshTokenService;
